Add support for `excludeFunctions`.
This is a proposal PR for supporting extra signaling/metadata in poko annotations for `excludeFunctions` (and leaving a toe-hold for future such configurations).

The intention behind this is that some types, such as Kotlin function types, are known not to have true equality and should be ignored. This especially comes to a head in compose, where capturing lambdas for objects with state become a major headache (https://issuetracker.google.com/issues/256100927).

This is a proposal for how this could look, where both the first-party `@Poko` annotation and any other custom ones could implement this. The idea is that a custom annotation could offer this with its own default. We could also even remove this from `@Poko` and only leave it as an optional control for custom annotations.

If this looks acceptable, I'll proceed with adding tests too.