Migrate GenericArrayHolder tests (#203)
The existing approach for creating an "&&" operation failed during lowering to native. The new approach of using a two-branch "when" was inspired by what the Compose compiler is doing for this. I could not find any other instances of this operation in the regular Kotlin compiler and this seems to work well. The generated code for this is okay (I checked JS
and JVM). I did not check before, so I have nothing to compare it to, but it's correct and so optimizing the generated code will become a future problem.