Migrate ArrayHolder test
The hashCode logic for primitive arrays was broken in that it would return the data class function to compute the hashCode. This should have been the instance-based function, which should have failed the JVM tests. Somehow, this only failed in JS and not JVM or native. Updated to the logic to use the correct contentHashCode function instead.